Make fish_key_reader available as a builtin
This brings us closer to making fish available as a single file.
